requirements.txt
# Dependencias esenciales para Panel Bet IA
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0

# Machine Learning básico
scikit-learn>=1.3.0

# Utilidades
python-dateutil>=2.8.0
